% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Measurement System Analysis Data}
\format{
\subsection{\code{data}}{

A data frame with 27 rows and 3 columns:
\describe{
\item{SN}{Serial Number}
\item{Operator}{Operator or Measurement Technician}
\item{Measure}{Measured Value}
...
}
}
}
\usage{
data
}
\description{
A collection of sample data to analyze for reproducibility and repeatability measurement error.
}
\keyword{datasets}
