% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_error.R
\name{squared_error}
\alias{squared_error}
\title{Calculate squared_error}
\usage{
squared_error(meas)
}
\arguments{
\item{meas}{A numeric vector}
}
\value{
A number.
}
\description{
Calculate squared_error
}
\examples{
squared_error(c(1.001, 1.003, 1.001))
squared_error(c(-.005, .003, .001))
}
