% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar_r_varcomps.R
\name{xbar_repeat}
\alias{xbar_repeat}
\title{XBAR & R Variance Component Calculations}
\usage{
xbar_repeat(data, part, operator, meas)
}
\arguments{
\item{meas}{}
}
\description{
XBAR & R Variance Component Calculations
}
