% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grr_comps.R
\name{grr_calc}
\alias{grr_calc}
\title{GRR Calculations}
\usage{
grr_calc(data, part, operator, meas, LSL = NULL, USL = NULL, method = "anova")
}
\arguments{
\item{data}{An R dataframe or tible.}

\item{part}{The column in data specifying the unique ID of the part being measured}

\item{operator}{A column in data specifying the operator for the recorded measurement}

\item{meas}{A column in data where the measurement value is recorded.}

\item{LSL}{A number to specify the lower specification limit.}

\item{USL}{A number to specify the upper specification limit.}

\item{method}{A string specifying 'anova' or 'xbar_r'}
}
\value{
A list of two dataframes: VarianceComponents and GageEval
}
\description{
GRR Calculations
}
\examples{

data = data.frame(
SN = c(
'SerialNumber_01',
'SerialNumber_01',
'SerialNumber_01',
'SerialNumber_01',
'SerialNumber_01',
'SerialNumber_01'),

Operator = c(
'Operator_01',
'Operator_01',
'Operator_01',
'Operator_02',
'Operator_02',
'Operator_02'),

Measure = c(
0.0173,
0.0151,
0.0173,
0.0163,
0.0155,
0.0175))

grr_calc(data, part = SN, operator = Operator, meas = Measure, LSL = 0, USL = .040, method = 'xbar_r')
}
