% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_varcomps.R
\name{ss_calcs}
\alias{ss_calcs}
\title{ANOVA Variance Components}
\usage{
ss_calcs(data, part, operator, meas)
}
\arguments{
\item{data}{An R dataframe or tible.}

\item{part}{The column in data specifying the unique ID of the part being measured}

\item{operator}{A column in data specifying the operator for the recorded measurement}

\item{meas}{A column in data where the measurement value is recorded.}
}
\value{
A list of numeric values for repeatability, reproducibility, total GRR, part-to-part, and total variance components.
}
\description{
ANOVA Variance Components
}
