% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xbar_r_varcomps.R
\name{xbar_repeat}
\alias{xbar_repeat}
\title{XBAR & R Variance Component Calculations}
\usage{
xbar_repeat(data, part, operator, meas)
}
\arguments{
\item{data}{An R dataframe or tible.}

\item{part}{The column in data specifying the unique ID of the part being measured}

\item{operator}{A column in data specifying the operator for the recorded measurement}

\item{meas}{A column in data where the measurement value is recorded.}
}
\value{
A list of numeric values for repeatability, reproducibility, total GRR, part-to-part, and total variance components.
}
\description{
XBAR & R Variance Component Calculations
}
